// web/src/api/generated/model/getInstallmentSeries200InstallmentsItem.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * HouseApp API
 * API for HouseApp
 * OpenAPI spec version: 2.3.0
 */
import type { GetInstallmentSeries200InstallmentsItemPaidAmount } from "./getInstallmentSeries200InstallmentsItemPaidAmount";
import type { GetInstallmentSeries200InstallmentsItemStatus } from "./getInstallmentSeries200InstallmentsItemStatus";

export type GetInstallmentSeries200InstallmentsItem = {
  id: string;
  installmentNumber: number;
  date: string;
  amount: string;
  paidAmount: GetInstallmentSeries200InstallmentsItemPaidAmount;
  remaining: string;
  status: GetInstallmentSeries200InstallmentsItemStatus;
};